+++ lib/features/offeredfood/data/mapper/offered_food_mapper.dart
@@ -1,3 +1,4 @@
+import 'package:zachranobed/enums/food_category.dart';
 import 'package:zachranobed/features/offeredfood/domain/model/food_date_time.dart';
 import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart';
 import 'package:zachranobed/models/dto/delivery_dto.dart';
@@ -8,6 +9,14 @@ import 'package:zachranobed/models/dto/meal_dto.dart';
 extension OfferedFoodMapper on MealDetailDto {
   /// Maps DTOs to domain representation.
   OfferedFood toDomain(DeliveryDto delivery, MealDto meal, String boxType) {
+    // The [preparedAt] field is set only for "cooled" meals, otherwise is set
+    // to null. This is because the [preparedAt] field is only relevant for
+    // cooled food categories.
+    FoodDateTime? preparedAt;
+    if (foodCategoryType == FoodCategoryType.cooled.name) {
+      preparedAt = _getFoodDateTime(meal.preparedAt);
+    }
+
     return OfferedFood(
       id: id,
       date: delivery.deliveryDate,
@@ -16,6 +25,7 @@ extension OfferedFoodMapper on MealDetailDto {
       allergens: allergens,
       numberOfServings: meal.count,
       boxType: boxType,
+      preparedAt: preparedAt,
       consumeBy: _getFoodDateTime(meal.consumeBy),
       donorId: delivery.donorId,
       recipientId: delivery.recipientId,