+++ lib/features/offeredfood/domain/model/food_info.dart
@@ -2,6 +2,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:uuid/uuid.dart';
 import 'package:zachranobed/common/helper_service.dart';
+import 'package:zachranobed/enums/food_category.dart';
+import 'package:zachranobed/features/foodboxes/domain/model/food_box_type.dart';
 import 'package:zachranobed/features/foodboxes/domain/repository/food_box_repository.dart';
 import 'package:zachranobed/features/offeredfood/domain/model/food_date_time.dart';
 
@@ -17,20 +19,22 @@ class FoodInfo with _$FoodInfo {
     required String id, // The UUID identifier
     String? dishName,
     List<String>? allergens,
-    String? foodCategory,
+    FoodCategory? foodCategory,
     int? numberOfServings,
     int? numberOfBoxes,
-    String? foodBoxId,
+    FoodBoxType? foodBoxType,
+    FoodDateTime? preparedAt,
     FoodDateTime? consumeBy,
   }) = _FoodInfo;
 
   factory FoodInfo.withUuid({
     String? dishName,
     List<String>? allergens,
-    String? foodCategory,
+    FoodCategory? foodCategory,
     int? numberOfServings,
     int? numberOfBoxes,
-    String? foodBoxId,
+    FoodBoxType? foodBoxType,
+    FoodDateTime? preparedAt,
     FoodDateTime? consumeBy,
   }) {
     return FoodInfo(
@@ -41,13 +45,35 @@ class FoodInfo with _$FoodInfo {
       foodCategory: foodCategory,
       numberOfServings: numberOfServings,
       numberOfBoxes: numberOfBoxes,
-      foodBoxId: foodBoxId,
+      foodBoxType: foodBoxType,
+      preparedAt: preparedAt,
       consumeBy: consumeBy,
     );
   }
 }
 
-extension FoodInfoExtension on List<FoodInfo> {
+extension FoodInfoExtension on FoodInfo {
+  /// Creates a copy of this [FoodInfo] object with the given [foodCategory] and
+  /// updates the [preparedAt] field accordingly.
+  ///
+  /// If the [foodCategory] type is not [FoodCategoryType.cooled], the
+  /// [preparedAt] field will be set to null. This is because the [preparedAt]
+  /// field is only relevant for cooled food categories.
+  ///
+  /// Returns a new [FoodInfo] object with the updated values.
+  FoodInfo copyWithFoodCategory(FoodCategory? foodCategory) {
+    FoodDateTime? newPreparedAt = preparedAt;
+    if (foodCategory?.type != FoodCategoryType.cooled) {
+      newPreparedAt = null;
+    }
+    return copyWith(
+      foodCategory: foodCategory,
+      preparedAt: newPreparedAt,
+    );
+  }
+}
+
+extension FoodInfoListExtension on List<FoodInfo> {
   ///Extension on List<FoodInfo> to verify the available box count.
   Future<bool> verifyAvailableBoxCount(
       BuildContext context, FoodBoxRepository foodBoxRepository) async {
@@ -58,7 +84,7 @@ extension FoodInfoExtension on List<FoodInfo> {
 
     final requiredBoxes = <String, int>{};
     for (final foodInfo in this) {
-      final foodBoxId = foodInfo.foodBoxId;
+      final foodBoxId = foodInfo.foodBoxType?.id;
       if (foodBoxId == null) {
         continue;
       }