+++ lib/features/offeredfood/domain/model/offered_food.dart
@@ -17,6 +17,7 @@ class OfferedFood with _$OfferedFood {
     required List<String> allergens,
     required int numberOfServings,
     required String boxType,
+    required FoodDateTime? preparedAt,
     required FoodDateTime consumeBy,
     required String donorId,
     required String recipientId,