+++ lib/models/dto/meal_dto.g.dart
@@ -9,6 +9,8 @@ part of 'meal_dto.dart';
 MealDto _$MealDtoFromJson(Map<String, dynamic> json) => MealDto(
       mealId: json['mealId'] as String,
       count: (json['count'] as num).toInt(),
+      preparedAt: _$JsonConverterFromJson<Timestamp, DateTime>(
+          json['preparedAt'], const TimestampConverter().fromJson),
       consumeBy: _$JsonConverterFromJson<Timestamp, DateTime>(
           json['consumeBy'], const TimestampConverter().fromJson),
       foodBoxId: json['foodBoxId'] as String,
@@ -18,6 +20,8 @@ MealDto _$MealDtoFromJson(Map<String, dynamic> json) => MealDto(
 Map<String, dynamic> _$MealDtoToJson(MealDto instance) => <String, dynamic>{
       'mealId': instance.mealId,
       'count': instance.count,
+      'preparedAt': _$JsonConverterToJson<Timestamp, DateTime>(
+          instance.preparedAt, const TimestampConverter().toJson),
       'consumeBy': _$JsonConverterToJson<Timestamp, DateTime>(
           instance.consumeBy, const TimestampConverter().toJson),
       'foodBoxId': instance.foodBoxId,